defense/CIDER/requirements.txt
accelerate==0.21.0
bitsandbytes
decord==0.6.0
einops==0.6.1
httpx
iopath
joypy
matplotlib==3.8.4
numpy==1.23.5
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.550.52
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
omegaconf
openai
opencv-python
pandas
peft==0.2.0
pillow
scikit-image
seaborn
sentencepiece
tiktoken==0.6.0
timm==0.6.13
torch==2.4.1
torchaudio
torchdata
torchtext
torchvision
tqdm==4.66.2
transformers==4.37.2
transformers-stream-generator==0.0.5
visual_genome
wandb
webdataset